Restructure Event entity
